added auth for deleting task

+++ src/main/java/taskTracker/controller/TaskController.java
@@ -3,6 +3,7 @@ package taskTracker.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import taskTracker.exception.NoAccessToGroupException;
 import taskTracker.model.AppUser;
@@ -41,7 +42,18 @@ public class TaskController {
 
   @DeleteMapping("/{taskId}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  public void deleteTask(@PathVariable("taskId") Long id) {
+  @Transactional
+  public void deleteTask(@PathVariable("taskId") Long id, Authentication authentication) {
+    AppUserDetails appUserDetails = (AppUserDetails) authentication.getPrincipal();
+    AppUser appUser = appUserDetails.getAppUser();
+
+    Task task = taskService.findTaskWithOwnerAndMembers(id);
+
+    if (appUser.getId() != task.getTaskGroup().getOwner().getId() &&
+        task.getTaskGroup().getMembers().stream().noneMatch(a -> a.getId() == appUser.getId())) {
+      throw new NoAccessToGroupException(appUser.getId());
+    }
+
     taskService.deleteTask(id);
   }
 }
